+++ src/validate/__tests__/utils.unit.js
@@ -1,5 +1,5 @@
+import { addDepreciation, serversExist } from '../utils';
 import assert from 'assert';
-import { serversExist } from '../utils';
 
 describe('validator utils', () => {
   describe('serversExist', () => {
@@ -10,6 +10,21 @@ describe('validator utils', () => {
       const expectedLength = 1;
 
       assert(result.length === expectedLength);
+    });
+  });
+  describe('addDepreciation', () => {
+    it('should add a depreciation detail', () => {
+      const details = [];
+      const path = 'servers.test';
+      const reason = 'Use "testing" instead';
+      const link = 'http://google.com';
+
+      const [result] = addDepreciation(details, path, reason, link);
+
+      assert(result.type === 'depreciation');
+      assert(result.path === path);
+      assert(result.message.indexOf(reason) > -1);
+      assert(result.message.indexOf(link) > -1);
     });
   });
 });