+++ src/validate/index.js
@@ -1,5 +1,6 @@
 import * as utils from './utils';
 const { combineErrorDetails, VALIDATE_OPTIONS, improveErrors } = utils;
+import chalk from 'chalk';
 import joi from 'joi';
 import validateServer from './servers';
 
@@ -53,7 +54,49 @@ function validateAll(config) {
 }
 
 export default function validate(config) {
-  const errors = validateAll(config);
+  const errors = [];
+  const depreciations = [];
 
-  return errors.map(error => error.message);
+  validateAll(config).forEach(problem => {
+    if (problem.type === 'depreciation') {
+      depreciations.push(problem);
+    } else {
+      errors.push(problem);
+    }
+  });
+
+  return {
+    errors: errors.map(error => error.message),
+    depreciations: depreciations.map(depreciation => depreciation.message)
+  };
+}
+
+export function showErrors(errors) {
+  const lines = [];
+  const plural = errors.length > 1 ? 's' : '';
+
+  lines.push(`${errors.length} Validation Error${plural}`);
+
+  errors.forEach(error => {
+    lines.push(`  - ${error}`);
+  });
+
+  lines.push('');
+
+  console.log(chalk.red(lines.join('\n')));
+}
+
+export function showDepreciations(depreciations) {
+  const lines = [];
+  const plural = depreciations.length > 1 ? 's' : '';
+
+  lines.push(`${depreciations.length} Depreciation${plural}`);
+
+  depreciations.forEach(depreciation => {
+    lines.push(`  - ${depreciation}`);
+  });
+
+  lines.push('');
+
+  console.log(chalk.yellow(lines.join('\n')));
 }